refactor(key-management): extract repair dialog hooks
